requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
Pillow==9.0.0
sqlparse==0.4.2
tzdata==2021.5
